requirements.txt
skillcorner
psycopg2-binary
python-dotenv
